+++ src/GreenWallet/Tools.h
@@ -38,7 +38,7 @@ namespace Tools {
 
   uint64_t getDivisor();
 
-  uint64_t getScanHeight();
+  uint32_t getScanHeight();
 
   template <typename T, typename Function>
   std::vector<T> filter(std::vector<T> input, Function predicate)
