+++ src/GreenWallet/Transfer.h
@@ -16,7 +16,7 @@ void transfer(std::shared_ptr<WalletInfo> walletInfo, uint32_t height,
 
 void doTransfer(std::string address, uint64_t amount, uint64_t fee,
                 std::string extra, std::shared_ptr<WalletInfo> walletInfo,
-                uint32_t height, uint64_t mixin = WalletConfig::defaultMixin,
+                uint32_t height, uint16_t mixin = WalletConfig::defaultMixin,
                 std::string nodeAddress = std::string(), uint64_t nodeFee = 0);
 
 void sendMultipleTransactions(CryptoNote::WalletGreen &wallet,
